Moved the long array test out of the world test logic
